feat: Add Color and Type enum in Piece
Deleted all final constants in Piece class.
Added Color and Type enum in Piece class to substitute the constants.
Changed all parameters and codes which are dependent on the constants to use enum.

Added test codes in PieceTest class to test the methods getWhiteRepresentation() and getBlackRepresentation() in Piece::Type enum.