feat: Add calculate point logic
Added calculatePoint() and findPiecesByColor() in Board class to calculate each point of piece.
Added findPiecesByColor in Rank class to pick piece by color in a Rank.
Added double point in Type enum to allocate the point for each piece.
Added methods related to matching color and type to verify pieces.